mageia: Fallback message when display manager fails.
Provide a unit for displaying a message to the user when the
display manager fails. This can be when no dm application is
installed or when it exits with a failure code.

Much like the prefdm script, the actual message is included
in the initscripts package.

https://bugs.mageia.org/show_bug.cgi?id=4769

v2: Use agetty rather than mingetty + quit plymouth + don't wait for
    graphical.target
v3: Use Type=idle now available upstream
v4: Add back the prefdm.service script removed in v188
v5: Remove Conflicts on plymouth-quit.service, as plymouth quits
    reliably and gracefully now through systemd.