systemctl: Do not attempt native calls for enable/disable if sysvinit handleds all units.
If you have sysvinit compat enabled, it might deal with all the passed
in unit names leaving only a null termitated strv structure.

This cannot be iterated (sensibly) and thus we should just bail out
and not bother calling various native methods.

This should prevent a bogus warning regarding the lack of [Install]
sections in systemd units when enabling a sysvinit unit.