Add path to locale search.
Mageia seems to store it's locales in /usr/share/locale (and also seems
to copy it to /etc/locale??) so add in /usr/share/locale in addition to
/usr/lib/locale. We should investigate if we 'force' the share path and
make steps to move it if it's a pointless deviation.

We should also find out why we have /etc/locale/ directory tree populated.